refactor docs code + fix wrong enum summary
