+++ Docs/Api/Enumerations/GamePadStick.md
@@ -1,7 +1,5 @@
 # `GamePadStick`
 
-Simulate moving a gamepad joystick
-
 ## Example Use
 `GamePadStick.Left`
