+++ Docs/Api/Enumerations/PressState.md
@@ -1,7 +1,5 @@
 # `PressState`
 
-Simulate pressing or releasing (or both) gamepad buttons.
-
 ## Example Use
 `PressState.Press`
