+++ Support/BuildMarkdownDocs/Util/CommandLineArgs.cs
@@ -1,0 +1,10 @@
+﻿namespace BuildMarkdownDocs.Util;
+
+internal class CommandLineArgs
+{
+    public string AssemblyDirectory { get; set; }
+    public string XmlFilesDirectory { get; set; }
+    public string OutputDirectory { get; set; }
+    public string Template { get; set; }
+    public string Filter { get; set; }
+}
